clean up old files and lint rules
