--- app/src/main/res/layout/content_onboarding_vpn_intro.xml
@@ -1,102 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?><!--
-  ~ Copyright (c) 2020 DuckDuckGo
-  ~
-  ~ Licensed under the Apache License, Version 2.0 (the "License");
-  ~ you may not use this file except in compliance with the License.
-  ~ You may obtain a copy of the License at
-  ~
-  ~     http://www.apache.org/licenses/LICENSE-2.0
-  ~
-  ~ Unless required by applicable law or agreed to in writing, software
-  ~ distributed under the License is distributed on an "AS IS" BASIS,
-  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  ~ See the License for the specific language governing permissions and
-  ~ limitations under the License.
-  -->
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-                                                   xmlns:app="http://schemas.android.com/apk/res-auto"
-                                                   xmlns:tools="http://schemas.android.com/tools"
-                                                   android:layout_width="match_parent"
-                                                   android:layout_height="match_parent"
-                                                   tools:context="com.duckduckgo.mobile.android.vpn.ui.onboarding.VpnOnboardingActivity">
-
-    <FrameLayout
-        android:id="@+id/onboarding_page_header"
-        android:layout_width="match_parent"
-        android:layout_height="280dp"
-        android:background="@color/light_blue"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent">
-
-        <ImageView
-            android:id="@+id/onboarding_page_image"
-            android:layout_marginTop="@dimen/keyline_7"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:contentDescription="@string/atp_OnboardingLogoDescription"
-            android:importantForAccessibility="no"
-            android:scaleType="centerInside"
-            app:srcCompat="@drawable/device_shield_onboarding_report"/>
-
-    </FrameLayout>
-
-    <TextView
-        android:id="@+id/onboarding_page_title"
-        android:textAppearance="@style/TextAppearance.DuckDuckGo.Headline2"
-        android:textSize="21sp"
-        android:textColor="@color/almostBlackDark"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="40dp"
-        android:paddingStart="50dp"
-        android:paddingEnd="50dp"
-        android:text="@string/atp_daxOnboardingIntroTitle"
-        android:textAlignment="center"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/onboarding_page_header" />
-
-    <TextView
-        android:id="@+id/onboarding_page_text"
-        android:textSize="16sp"
-        android:textColor="@color/gray90"
-        android:textAlignment="center"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="@dimen/keyline_2"
-        android:layout_marginStart="@dimen/keyline_5"
-        android:layout_marginEnd="@dimen/keyline_5"
-        android:drawablePadding="@dimen/keyline_2"
-        android:text="@string/atp_daxOnboardingIntroMessage"
-        app:drawableTopCompat="@drawable/ic_beta_pill"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/onboarding_page_title" />
-
-    <com.duckduckgo.mobile.android.ui.view.button.ButtonPrimaryRounded
-        android:id="@+id/onboarding_next_cta"
-        android:text="@string/atp_OnboardingContinue"
-        android:layout_width="0dp"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="@dimen/keyline_5"
-        android:layout_marginEnd="@dimen/keyline_5"
-        android:layout_marginBottom="@dimen/keyline_4"
-        app:layout_constraintBottom_toTopOf="@+id/onboarding_maybe_later"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"/>
-
-    <com.duckduckgo.mobile.android.ui.view.button.ButtonSecondaryRounded
-        android:id="@+id/onboarding_maybe_later"
-        android:layout_width="0dp"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="@dimen/keyline_5"
-        android:layout_marginEnd="@dimen/keyline_5"
-        android:layout_marginBottom="20dp"
-        android:text="@string/atp_OnboardingMaybeLater"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"/>
-
-
-</androidx.constraintlayout.widget.ConstraintLayout>
