--- app/src/main/res/layout/item_tracker_network_section_title.xml
@@ -1,60 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?><!--
-  ~ Copyright (c) 2022 DuckDuckGo
-  ~
-  ~ Licensed under the Apache License, Version 2.0 (the "License");
-  ~ you may not use this file except in compliance with the License.
-  ~ You may obtain a copy of the License at
-  ~
-  ~     http://www.apache.org/licenses/LICENSE-2.0
-  ~
-  ~ Unless required by applicable law or agreed to in writing, software
-  ~ distributed under the License is distributed on an "AS IS" BASIS,
-  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  ~ See the License for the specific language governing permissions and
-  ~ limitations under the License.
-  -->
-
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    android:orientation="vertical">
-
-    <com.duckduckgo.mobile.android.ui.view.divider.HorizontalDivider
-        android:id="@+id/trackersTopDivider"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="@dimen/keyline_5"
-        android:layout_marginEnd="@dimen/keyline_5"
-        android:layout_marginBottom="@dimen/keyline_5"
-        android:layout_marginTop="@dimen/keyline_3" />
-
-    <TextView
-        android:id="@+id/trackersSectionDescription"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:paddingStart="@dimen/keyline_5"
-        android:paddingEnd="@dimen/keyline_5"
-        android:textAlignment="center"
-        tools:text="We blocked the following third-party domains' tracking requests from loading." />
-
-    <TextView
-        android:id="@+id/trackersSectionLink"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:paddingStart="@dimen/keyline_5"
-        android:paddingTop="@dimen/keyline_1"
-        android:paddingEnd="@dimen/keyline_5"
-        android:textAlignment="center"
-        android:visibility="gone"
-        tools:text="Learn more"
-        tools:visibility="visible" />
-
-    <com.duckduckgo.mobile.android.ui.view.divider.HorizontalDivider
-        android:id="@+id/trackersBottomDivider"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="@dimen/keyline_5"
-        android:layout_marginTop="@dimen/keyline_5"
-        android:layout_marginEnd="@dimen/keyline_5"/>
-</LinearLayout>
