+++ app/src/main/res/layout/view_tab_switcher_button.xml
@@ -20,6 +20,7 @@
 
     <!-- This TextView doesn't migrate to DaxTextView, special case for the Tabs icon  -->
     <TextView
+            tools:ignore="MergeRootFrame"
         android:id="@+id/tabCount"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
