+++ lint-rules/src/test/java/com/duckduckgo/lint/ui/DeprecatedAndroidWidgetsUsedInXmlDetectorTest.kt
@@ -109,4 +109,190 @@ class DeprecatedAndroidWidgetsUsedInXmlDetectorTest {
             .run()
             .expectClean()
     }
+
+    @Test
+    fun whenAppCompatSwitchFoundThenFailWithError() {
+        lint()
+            .files(
+                TestFiles.xml(
+                    "res/layout/switch.xml",
+                    """
+                <android.support.design.widget.CoordinatorLayout
+                    xmlns:android="http://schemas.android.com/apk/res/android"
+                    xmlns:app="http://schemas.android.com/apk/res-auto"
+                    xmlns:tools="http://schemas.android.com/tools"
+                    android:layout_width="match_parent"
+                    android:layout_height="match_parent"
+                    android:background="#eeeeee">
+
+                  <android.support.design.bottomappbar.BottomAppBar
+                      android:id="@+id/bottom_app_bar"
+                      style="@style/Widget.MaterialComponents.BottomAppBar"
+                      android:layout_width="match_parent"
+                      android:layout_height="wrap_content"
+                      android:layout_gravity="bottom"
+                      app:navigationIcon="@drawable/ic_menu_black_24dp"/>
+
+                <androidx.appcompat.widget.SwitchCompat
+                    android:id="@+id/setAsDefaultBrowserSetting"
+                    style="@style/SettingsSwitch"
+                    android:paddingTop="14dp"
+                    android:paddingBottom="14dp"
+                    android:text="@string/settingsDefaultBrowserEnabled"
+                    android:theme="@style/SettingsSwitchTheme" />
+
+
+                </android.support.design.widget.CoordinatorLayout>
+            """
+                ).indented()
+            )
+            .issues(DeprecatedAndroidWidgetsUsedInXmlDetector.DEPRECATED_WIDGET_IN_XML)
+            .run()
+            .expect(
+                """
+                res/layout/switch.xml:17: Error: Always favor the use of the Design System Component SwitchView [AndroidSwitchInXml]
+                <androidx.appcompat.widget.SwitchCompat
+                 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+                1 errors, 0 warnings
+            """.trimMargin()
+            )
+    }
+
+    @Test
+    fun whenMaterialSwitchFoundThenFailWithError() {
+        lint()
+            .files(
+                TestFiles.xml(
+                    "res/layout/switch.xml",
+                    """
+                <android.support.design.widget.CoordinatorLayout
+                    xmlns:android="http://schemas.android.com/apk/res/android"
+                    xmlns:app="http://schemas.android.com/apk/res-auto"
+                    xmlns:tools="http://schemas.android.com/tools"
+                    android:layout_width="match_parent"
+                    android:layout_height="match_parent"
+                    android:background="#eeeeee">
+
+                  <android.support.design.bottomappbar.BottomAppBar
+                      android:id="@+id/bottom_app_bar"
+                      style="@style/Widget.MaterialComponents.BottomAppBar"
+                      android:layout_width="match_parent"
+                      android:layout_height="wrap_content"
+                      android:layout_gravity="bottom"
+                      app:navigationIcon="@drawable/ic_menu_black_24dp"/>
+
+                    <com.google.android.material.switchmaterial.SwitchMaterial
+                        android:id="@+id/enabledToggle"
+                        android:layout_width="wrap_content"
+                        android:layout_height="wrap_content"
+                        android:layout_marginStart="@dimen/keyline_4"
+                        android:layout_marginEnd="@dimen/keyline_4"
+                        android:gravity="center_vertical"
+                        app:layout_constraintBottom_toBottomOf="@id/enabledToggleLabel"
+                        app:layout_constraintEnd_toEndOf="parent"
+                        app:layout_constraintTop_toTopOf="@id/enabledToggleLabel" />
+
+                </android.support.design.widget.CoordinatorLayout>
+            """
+                ).indented()
+            )
+            .issues(DeprecatedAndroidWidgetsUsedInXmlDetector.DEPRECATED_WIDGET_IN_XML)
+            .run()
+            .expect(
+                """
+                res/layout/switch.xml:17: Error: Always favor the use of the Design System Component SwitchView [AndroidSwitchInXml]
+                    <com.google.android.material.switchmaterial.SwitchMaterial
+                     ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+                1 errors, 0 warnings      
+            """.trimMargin()
+            )
+    }
+
+    @Test
+    fun whenLegacySwitchFoundThenFailWithError() {
+        lint()
+            .files(
+                TestFiles.xml(
+                    "res/layout/switch.xml",
+                    """
+                <android.support.design.widget.CoordinatorLayout
+                    xmlns:android="http://schemas.android.com/apk/res/android"
+                    xmlns:app="http://schemas.android.com/apk/res-auto"
+                    xmlns:tools="http://schemas.android.com/tools"
+                    android:layout_width="match_parent"
+                    android:layout_height="match_parent"
+                    android:background="#eeeeee">
+
+                  <android.support.design.bottomappbar.BottomAppBar
+                      android:id="@+id/bottom_app_bar"
+                      style="@style/Widget.MaterialComponents.BottomAppBar"
+                      android:layout_width="match_parent"
+                      android:layout_height="wrap_content"
+                      android:layout_gravity="bottom"
+                      app:navigationIcon="@drawable/ic_menu_black_24dp"/>
+
+                <Switch
+                    android:id="@+id/setAsDefaultBrowserSetting"
+                    style="@style/SettingsSwitch"
+                    android:paddingTop="14dp"
+                    android:paddingBottom="14dp"
+                    android:text="@string/settingsDefaultBrowserEnabled"
+                    android:theme="@style/SettingsSwitchTheme" />
+
+
+                </android.support.design.widget.CoordinatorLayout>
+            """
+                ).indented()
+            )
+            .issues(DeprecatedAndroidWidgetsUsedInXmlDetector.DEPRECATED_WIDGET_IN_XML)
+            .run()
+            .expect(
+                """
+                res/layout/switch.xml:17: Error: Always favor the use of the Design System Component SwitchView [AndroidSwitchInXml]
+                <Switch
+                 ~~~~~~
+                1 errors, 0 warnings
+            """.trimMargin()
+            )
+    }
+
+    @Test
+    fun whenDesignSystemComponentSwitchFoundThenSucceed() {
+        lint()
+            .files(
+                TestFiles.xml(
+                    "res/layout/switch.xml",
+                    """
+                <android.support.design.widget.CoordinatorLayout
+                    xmlns:android="http://schemas.android.com/apk/res/android"
+                    xmlns:app="http://schemas.android.com/apk/res-auto"
+                    xmlns:tools="http://schemas.android.com/tools"
+                    android:layout_width="match_parent"
+                    android:layout_height="match_parent"
+                    android:background="#eeeeee">
+
+                  <android.support.design.bottomappbar.BottomAppBar
+                      android:id="@+id/bottom_app_bar"
+                      style="@style/Widget.MaterialComponents.BottomAppBar"
+                      android:layout_width="match_parent"
+                      android:layout_height="wrap_content"
+                      android:layout_gravity="bottom"
+                      app:navigationIcon="@drawable/ic_menu_black_24dp"/>
+                      
+                <com.duckduckgo.mobile.android.ui.view.listitem.OneLineListItemSwitch
+                      android:id="@+id/bottom_app_bar"
+                      android:layout_width="match_parent"
+                      android:layout_height="wrap_content"
+                      android:layout_gravity="bottom"
+                      app:navigationIcon="@drawable/ic_menu_black_24dp"/>
+
+                </android.support.design.widget.CoordinatorLayout>
+            """
+                ).indented()
+            )
+            .allowCompilationErrors()
+            .issues(DeprecatedAndroidWidgetsUsedInXmlDetector.DEPRECATED_WIDGET_IN_XML)
+            .run()
+            .expectClean()
+    }
 }
